Phase 1 Step 1: extract MicrostructureService (template step)
- quantcore/gateways/yfinance_gateway.py: YFinanceGateway (ticker_info with
  15s hard timeout, fast_info, expirations, option_chain) — all yfinance
  network access for the services layer.
- quantcore/services/microstructure.py: MicrostructureService with
  get_short_interest / get_dark_pool / get_bid_ask_spread, bodies moved
  verbatim from market_analysis_server.py; constructor injection of
  OhlcvRepository + YFinanceGateway.
- registry: wires gateway + service.
- market_analysis_server.py: now a thin adapter — three @mcp.tool()
  functions, one service call each, docstrings preserved verbatim.
- stock_price_server.py: cross-server import of the three decorated tools
  replaced with get_services().microstructure calls (fastmcp 3.2.3 returns
  the original function from @mcp.tool(), so the old cross-calls were plain
  delegation — behavior identical).

Parity-verified against pre-refactor baseline (AAPL, all three tools, all
fields identical). 30 unit tests green against the test DB; all 5 MCP
servers import; Flask boots 38 routes.
